Refined REST API contracts to be compliant

+++ src/main/java/com/coinpurse/web/services/impl/EventServicesImpl.java
@@ -1,7 +1,6 @@
 package com.coinpurse.web.services.impl;
 
 import com.coinpurse.web.domain.exceptions.ResourceNotFoundException;
-import com.coinpurse.web.exceptions.GlobalExceptionHandler;
 import com.coinpurse.web.model.Event;
 import com.coinpurse.web.model.Purse;
 import com.coinpurse.web.repository.EventRepository;
@@ -12,11 +11,11 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.time.temporal.ChronoUnit;
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
 import static com.coinpurse.web.constants.ErrorMessages.EVENT_NOT_FOUND;
+import static com.coinpurse.web.constants.ErrorMessages.PURSE_NOT_FOUND;
 
 @Service
 public class EventServicesImpl implements EventServices {
@@ -56,15 +55,19 @@ public class EventServicesImpl implements EventServices {
                 .orElseThrow(() -> new ResourceNotFoundException(EVENT_NOT_FOUND));
     }
 
-    // TODO: Should return 404 if not found
     @Override
-    public void updatePurse(Event event) {
-        eventRepository.save(event);
+    public Event updateEvent(Event event) {
+        if(event.getId() == null || eventRepository.existsById(event.getId())) {
+            throw new ResourceNotFoundException(EVENT_NOT_FOUND);
+        }
+        return eventRepository.save(event);
     }
 
-    // TODO: Should return 404 if not found, 204 if found
     @Override
     public void deleteEvent(Event event) {
+        if(event.getId() == null || eventRepository.existsById(event.getId())) {
+            throw new ResourceNotFoundException(EVENT_NOT_FOUND);
+        }
         eventRepository.deleteById(event.getId());
     }
 
@@ -83,7 +86,7 @@ public class EventServicesImpl implements EventServices {
     @Override
     public List<Event> getEventsByPurse(Purse purse) {
         List<Event> events = eventRepository.getEventsByPurse(purse);
-        Collections.sort(events, Comparator.comparing(Event::getDate));
+        events.sort(Comparator.comparing(Event::getDate));
 
         return events;
     }
